+++ src/elements/GlobalSet.php
@@ -207,6 +207,17 @@ class GlobalSet extends Element
     /**
      * @inheritdoc
      */
+    public function attributeLabels(): array
+    {
+        return [
+            'handle' => Craft::t('app', 'Handle'),
+            'name' => Craft::t('app', 'Name'),
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
     protected function defineRules(): array
     {
         $rules = parent::defineRules();