+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Fixed
 - Fixed the vertical alignment of condition rules’ sort handles. ([#11025](https://github.com/craftcms/cms/issues/11025))
+- Fixed a JavaScript error that occurred if `Craft.cp.$tabsList` or `Craft.cp.$tabs` were referenced.
 
 ## 4.0.0-RC1 - 2022-04-26